PATCH: [perl #112208]: Set utf8 flag on $! appropriately
This patch sets the utf8 flag on $! if the error string passes utf8
validity tests and has some bytes with the upper bit set.  (If none
have that bit set, is an ASCII string, and whether or not it is UTF-8 is
irrelevant.)  This is a heuristic that could fail, but as the reference
in the comments points out this is unlikely.

One can reasonably assume that a UTF-8 locale will return a UTF-8
result.  So another approach would be to look at that (but we wouldn't
want to turn the flag on for a purely ASCII string anyway, as that could
change the semantics from existing behavior by making the string follow
Unicode rules, whereas it didn't have to before.)  But we have no sure
way of knowing if we have a UTF-8 locale or not.  The string returned by
setlocale() is defined as opaque, and on WIndows at least it really is.
When a locale has messages outside ASCII, one could pretty reliably
see if it is a UTF-8 locale by examining for UTF-8 validity the return
of strerror() of every possible errno.  This could be done at setlocale
time.  But until the heuristic in this patch is shown to not be good
enough, I don't see the need to do this extra work.