Fix broken qr/ \N* /x
Under /x, the parsing was skipping the character following a \N (when
meaning [^\n]).  In fact the parse pointer was already pointing to the
first non-white-space character, so there was no need to advance it, and
then back off.  The problem was that the advance was potentially many
characters, and the back-off was always just one.