Remove useless if(PL_stashcache) checks
For some reason, various places checked PL_stashcache for truth before
using it. Yet it's always initialized to be a hash and other bits of
code appear to use it quite ignorant of whether it's NULL or otherwise.
This suggests that this is an remnant that had been cargo-culted around.