fix B::METHOP::first, B::METHOP::meth_sv
for METHOPs, op_first and op_meth_sv are in a union, so only
one can be valid a time. Decide which based on op type. Previously
it was interpreting an SV pointer as an OP pointer and vice versa.

Also, remove some vestigal code from Concise.pm that handled
method_named, in a branch no longer called for method ops